requirements.txt
google-cloud-datastore>=2.1.6
google-cloud-pubsub>=2.8.0
